requirements-payments.txt
# requirements-payments.txt
djangorestframework==3.14.0
requests==2.31.0
cryptography>=45.0.7,<47
python-decouple==3.8
pyotp==2.8.0
qrcode==7.4.2
reportlab==4.0.6
django-cors-headers==4.3.1
django-environ==0.10.0